+++ packages/solid/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Refactor `setError` and `validate` method
 - Fix focus bug at `setError` method
 
 ## v0.1.2 (October 21, 2022)